feat: add migration manager for older schemas
