+++ src/main/java/com/tatayless/sovereignty/database/DatabaseManager.java
@@ -11,6 +11,7 @@ import org.jooq.impl.DSL;
 import java.io.File;
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.locks.ReentrantLock;
 
 public class DatabaseManager {
@@ -31,6 +32,29 @@ public class DatabaseManager {
         setupDataSource();
         tableManager = new TableManager(plugin, configManager.isMySQL());
         createTablesIfNotExist();
+    }
+
+    /**
+     * Explicitly run database migrations.
+     * This is useful when you want to force migrations to run
+     * outside of the normal startup sequence.
+     */
+    public CompletableFuture<Boolean> runMigrations() {
+        CompletableFuture<Boolean> future = new CompletableFuture<>();
+
+        plugin.getServer().getScheduler().runTaskAsynchronously(plugin, () -> {
+            try (Connection conn = getConnection()) {
+                DSLContext context = createContextSafe(conn);
+                boolean success = tableManager.getMigrationManager().applyMigrations(conn, context);
+                future.complete(success);
+            } catch (SQLException e) {
+                plugin.getLogger().severe("Failed to run migrations: " + e.getMessage());
+                e.printStackTrace();
+                future.completeExceptionally(e);
+            }
+        });
+
+        return future;
     }
 
     private void setupDataSource() {
